Allow ignoring of missing workbooks
During the evaluation of a formula, if POI cannot resolve a referenced
spreadsheet it will throw:

CollaboratingWorkbooksEnvironment.WorkbookNotFoundException

Sometimes this is not the desired behaviour and POI provides a way of
telling the evaluator to ignore these failures and just returned the
cached value for the cell using `setIgnoreMissingWorkbooks`.

In order to expose this behaviour we have used a dynamic var so that we
can wrap the processing of a whole spreadsheet in a binding, rather than
passing properties through to each call of `read-cell`.